All resized and checked, need to free and one more function to go
